feat: add temp setter for badgertree

+++ interfaces/solidly/IBaseV1Pair.sol
@@ -1,3 +1,4 @@
+// TODO: fix this
 // SPDX-License-Identifier: MIT
 
 pragma solidity ^0.6.0;
@@ -7,12 +8,40 @@ import "../erc20/IERC20.sol";
 
 interface IBaseV1Pair is IERC20 {
     function token0() external view returns (address);
+
     function token1() external view returns (address);
 
-    function permit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
-    function burn(address to) external returns (uint amount0, uint amount1);
-    function mint(address to) external returns (uint liquidity);
-    function getReserves() external view returns (uint112 _reserve0, uint112 _reserve1, uint32 _blockTimestampLast);
-    function getAmountOut(uint, address) external view returns (uint);
+    function permit(
+        address owner,
+        address spender,
+        uint256 value,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external;
+
+    function swap(
+        uint256 amount0Out,
+        uint256 amount1Out,
+        address to,
+        bytes calldata data
+    ) external;
+
+    function burn(address to)
+        external
+        returns (uint256 amount0, uint256 amount1);
+
+    function mint(address to) external returns (uint256 liquidity);
+
+    function getReserves()
+        external
+        view
+        returns (
+            uint112 _reserve0,
+            uint112 _reserve1,
+            uint32 _blockTimestampLast
+        );
+
+    function getAmountOut(uint256, address) external view returns (uint256);
 }
